Follow-up r95155 - Make PagedTiffHandler use the new method to make sure the Exif support takes endian-ness of the tiff file into account
(needed because php's exif plugin doesn't do it for us... Only affects a few obscure props)

Also while I am here - change the FormatExif call to FormatMetadata (FormatExif is deprecated).
Also remove the htmlspecialchars - it doesn't make sense at this point, the result of that method is
fed through the parser in ImagePage.php and it will screw up formatting of Exif properties that
have multiple values (are arrays) and should be displayed as a list.