Check $wgMaxImageArea before downloading file asset from swift
Transfering a hundreds of megabytes file around the servers just
to throw it away seems rather silly. We should check for the
file is too big condition before fetching the file.

I suspect this is probably part of the problem with recent
gwtoolset-thumbnailing scalars load issues (bug 65217).

This also brings the file to big error inline with how this error
for other file types are displayed. I left the max area check
in the doTransform method just to be safe, but the new code
should cause an error before that code is reached.

Bug: 62306
Change-Id: Id858d6c0afeb30b8baf3731a566b32450e54d659