Rename tests/selenium/ folder to "phpunit-selenium"
There is a convention for Selenium tests (written in JavaScript) to
be in this directory. What this directory contains is different.
There is a PHPUnit extension required to run this tests. I suspect
they have not been run since 2010 when they have been written. We
might as well remove this folder.

Bug: T199116
Change-Id: I11dbc2d362f05a4c70cfbe8d2e1dcad52c57f383
Needed-By: I8e77ffab860e488d08a07ca8cbc76eefa1bdf0ca