Fix index error in PagedTiffHandler::normaliseParams
I found this error via logstash.wikimedia.org. There is
apparently no guarantee this field exists. I think this is
fine. This line of code decides if the preview format is JPEG
(default) or PNG (only when there is transparency in the TIFF).
I think we can assume JPEG is fine when there is no info about
an alpha channel.

Change-Id: I2f74d3ccde1f93a05b300e0494cfe74b3e1d1c6c