Allow proper caching of PagedTiffImage and its metadata
When PagedTiffHandler::verifyUpload is called, we always generate
a new PagedTiffImage, so the metadata are retrieved again.

For multi-staged uploads, and in particular for very large files, this
results in multiple wasteful shellouts or RPC calls (if shellbox is
used), wasting cpu time and slowing the whole process down.

In order to account for partial uploads verifications, the metadata of
the image are cleaned out if verification fails.

Bug: T292322
Change-Id: I5ea6f6668786fba727cd3a79c5244a9dc344b209